Add tests for tokens
