+++ src/bpp/__init__.py
@@ -2,6 +2,6 @@
 
 from .exceptions import BrainfuckSyntaxError
 from .interpreter import Interpreter
-from .tokens import Token
+from .tokens import Token, tokenize
 
-__all__ = ["Interpreter", "Token", "BrainfuckSyntaxError"]
+__all__ = ["Interpreter", "Token", "BrainfuckSyntaxError", "tokenize"]
